Fixes issue with sorting in PHP7
This commit alters the sort-algorithm used for testing which before was
based on the position of the entry in the array and is now solely based
on the value itself.

Due to changes in PHP7 sorting based on the key of the value was not
working in PHP7 as it was in PHP5 so there was need for a change.

Nowe the values are sorted by the number of "1" in the binary
representation of the value. If that is equals the natural sorting
algorithm is used.