Fixes BC-Issue due to removed error-handler
This commit fixes a BC-Issue caused by removing the
Zend\StdLib\ErrorHandler and replacing it with a class that does
nothing. Out of the Box that caused some functions to throw errors
instead of the expected Exceptions.

The replacement-class for the Zend\StdLib\ErrorHandler now sets (and
removes) an empty errorHandler.

Thanks to @sgehrig for mentioning and reaching out!